Added new endpoint `GetMapping` for fetching a single mapping.
Added new endpoint `GetAllMapping` for fetching all mappings belonging to the logged in identity.

Session agent now stores the reference to the operation it is handling.  This makes it possible to include a `Stop` method used by external callers to stop the agent and close the operation's wait channel.  The agent will also include an ErrAgentBusy to a new operation and close it's wait channel if it is (the agent) busy.

Added tests for new endpoints and fixed failing tests.